lexer/lexer_test.go
package lexer

import (
	"go-interpreter-demo/token"
	"testing"
)

func TestNextTokenSimple(t *testing.T) {
	sourceCode := `=+(){},;`

	tests := []struct {
		expectType      token.Type
		expectedLiteral string
	}{
		{token.ASSIGN, "="},
		{token.PLUS, "+"},
		{token.LPAREN, "("},
		{token.RPAREN, ")"},
		{token.LBRACE, "{"},
		{token.RBRACE, "}"},
		{token.COMMA, ","},
		{token.SEMICOLON, ";"},
	}

	l := New(sourceCode)

	for i, tt := range tests {
		tok := l.NextToken()

		if tok.Type != tt.expectType {
			t.Fatalf("test[%d] wrong token type %q, expected %q", i, tok.Type, tt.expectType)
		}

		if tok.Literal != tt.expectedLiteral {
			t.Fatalf("test[%d] wrong token literal %q, expected %q", i, tok.Literal, tt.expectedLiteral)
		}
	}
}

func TestNextToken(t *testing.T) {
	sourceCode := `let five = 5;
let ten = 10;

let add = fn(x, y) {
	x + y;
};

let result = add(five, ten);
`

	tests := []struct {
		expectType      token.Type
		expectedLiteral string
	}{
		{token.LET, "let"},
		{token.IDENT, "five"},
		{token.ASSIGN, "="},
		{token.INT, "5"},
		{token.SEMICOLON, ";"},
		{token.LET, "let"},
		{token.IDENT, "ten"},
		{token.ASSIGN, "="},
		{token.INT, "10"},
		{token.SEMICOLON, ";"},
		{token.LET, "let"},
		{token.IDENT, "add"},
		{token.ASSIGN, "="},
		{token.FUNCTION, "fn"},
		{token.LPAREN, "("},
		{token.IDENT, "x"},
		{token.COMMA, ","},
		{token.IDENT, "y"},
		{token.RPAREN, ")"},
		{token.LBRACE, "{"},
		{token.IDENT, "x"},
		{token.PLUS, "+"},
		{token.IDENT, "y"},
		{token.SEMICOLON, ";"},
		{token.RBRACE, "}"},
		{token.SEMICOLON, ";"},
		{token.LET, "let"},
		{token.IDENT, "result"},
		{token.ASSIGN, "="},
		{token.IDENT, "add"},
		{token.LPAREN, "("},
		{token.IDENT, "five"},
		{token.COMMA, ","},
		{token.IDENT, "ten"},
		{token.RPAREN, ")"},
		{token.SEMICOLON, ";"},
		{token.EOF, ""},
	}

	l := New(sourceCode)

	for i, tt := range tests {
		tok := l.NextToken()

		if tok.Type != tt.expectType {
			t.Fatalf("test[%d] - tokentype wrong[%q,%q], expected[%q]", i, tok.Type, tok.Literal, tt.expectType)
		}

		if tok.Literal != tt.expectedLiteral {
			t.Fatalf("test[%d] - tokenliteral wrong[%q], expected[%q]", i, tok.Literal, tt.expectedLiteral)
		}
	}
}
